+++ scripts/delete-branch.sh
@@ -27,9 +27,11 @@ function check_for_required_env_vars() {
 cleanup_dotenv() {
     local env_file=$1
 
-    # Use awk to remove empty lines, but retain the last empty line if it exists
-    awk 'NF && !found {print; found=1; next} NF {print}' "$env_file" > temp_env_file
-    mv temp_env_file "$env_file"
+    if [ -f .env ]; then
+          # Use awk to remove empty lines, but retain the last empty line if it exists
+        awk 'NF && !found {print; found=1; next} NF {print}' "$env_file" > temp_env_file
+        mv temp_env_file "$env_file"
+    fi
 }
 
 function branch_name_from_git() {
